refactoring, rename dvm-api -> dvm-net
