+++ net/src/lib.rs
@@ -13,26 +13,19 @@ pub mod serve;
 #[cfg(any(unix, macos))]
 mod unix;
 
+#[cfg(windows)]
+mod win;
+
 pub mod transport {
     #[cfg(any(unix, macos))]
     pub use super::unix::*;
 
     #[cfg(windows)]
-    compile_error!("windows platform is not supported");
+    pub use super::win::*;
 }
 
 pub mod prelude {
-    //! The DVM-NET Prelude
-    //!
-    //! The purpose of this module is to give imports-set for many common cases
-    //! by adding a glob import to the top:
-    //!
-    //! ```
-    //! # #![allow(unused_imports)]
-    //! use dvm_net::prelude::*;
-    //! ```
-    //!
-    //! Also contains reshared `TryInto` and `TryFrom` std traits.
+    // pub use dvm_api as api;
 
     pub use crate::serve::*;
     pub use crate::endpoint::*;
